Grouping BaseUnit tests
