+++ UnitTests/BaseUnit/Extensionstests.cs
@@ -2,10 +2,10 @@
 using EngineeringUnits.Units;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace UnitTests;
+namespace UnitTests.BaseUnit;
 
 [TestClass()]
-public class BaseUnitExtensionstests
+public class Extensionstests
 {
 
     [TestMethod()]
